+++ ChangeLog.md
@@ -6,6 +6,10 @@
 
 - Added a `diagnostics` CLI command to print out useful environmental
   information. ([#210](https://github.com/davep/rogallo/pull/210))
+- Added support for the `finger` protocol.
+  ([#211](https://github.com/davep/rogallo/pull/211))
+- Added `!finger <user[@host]>` as a command line command.
+  ([#211](https://github.com/davep/rogallo/pull/211))
 
 ## v1.0.0